Add ID "four" and styles for ex 02
I got the page to look the same (with some different colors, but the
correct properties). I chose to group the 3rd line class with the 4th
line ID, as the 3rd line's class was just used to change the font to
the same size as the 4th line, but I like the solution method more,
where they created a class that would be used to change a font size, and
which could then be applied to any element that would need to be that
size. This makes WAY more sense, and feels more reusable/versatile.