Add classes and styles for exercies 03
Admittedly, I misread the instructions, and set up a separate style for
the font size, as done in the previous exercise (in which I utilized
the grouping to be practiced here). I changed my file to be more
in line with the instructions. I also used different colors, using IDLE
to randomly generate numbers in both decimal and hex to use all three
types of color selector - Hex, RGB, and HSL.