Fix 2nd line and buttons for Exercise 06
.para was listed after .small-para in the css, and since the 2nd line
has both classes (and both have the same specificity), .para overrode
.small-para. Moving .small-para after .para fixed this.

The buttons share the button class, but each have their own unique
class. I created a .cancel class and put everything from .button in
there except for the font size, as that is to be the same for both. I
left the font size in a .button class. I think I could have used the
IDs for this, too, or added the font size to .button.confirm, and
changed .button to .button.cancel.