Set button as a block element
I suppose the button could have been a separate element, but that
doesn't feel like it's in the spirit of the exercise. I changed it's
display to block to get it to wrap, added the same padding as all the
other elements, and set its margins to auto to center it.