Change styles to correct values per readme
As it turns out, not all of the padding was exactly the same (and my
guesstimate of margin/padding of 10px was off by 2px - I'm not ashamed
by that). I adjusted all of the 10px padding to 8px, set the content
section to have 16px padding on top and bottom, and set the left and
right padding of the button to 24px. I also set the title font to the
correct size of 16px.