Change header and ul to flex containers
I used margin-left and margin-right both set to auto to center the logo
in the header flex container. The solution uses justify-content:
space-between, and no margins, which is a simpler/cleaner method. I
think I shied away from that because it is what I'd initially tried,
but I had other things wrong, so it didn't work.

It took me a while to realize that the ul had a padding that had to be
stripped by setting padding to 0. Solution adds margin: 0 to the ul as
well, but I don't think it does anything.