Style header to complete Exercise 03
I love that there is more than one way to do something in coding. I
chose to take advantage of auto margins and didn't create any new
classes. The solution creates two classes, .left and .right to wrap
the left-hand and right-hand elements in. Setting the justification on
the header container to space-between causes them to shift to either
end of the box. Then, they set those as flex containers, aligned their
items to center and added a gap. To make everything else work, they
set the ul on the left as a flex container (not sure why they wouldn't
have used the class selector - seems like their method could cause
problems if they use a ul later in the page) and set the cross-axis
alignment and gap.

Honestly, I like my solution more. I added flex properties to the
header and to the .links class (the ul), using margin-right: auto to
force the notification and profile image all the way to the right. I
can't imagine the links class would be used later in the page, and so
this shouldn't cause conflicts later, and it doesn't involve creating
another div, keeping the code in a cleaner, more readable state, at
least in my opinion.