Update CSS and HTML to complete Exercise 04
I got this one mostly correct. One tiny change I made after reading the
solution was to move text-align: center into the body css, as it
affects all of the body elements. I had added it to both the title and
the text fields.

As in the solution, I created two new classes: a container for all of
the plants/info, and one to apply to each plant/info combo. Where I
diverged (and perhaps unnecessarily) is that I set the flex direction
of these inner plant classes to column, so that each image would stack
on top of it's text. This isn't technically necessary as the default
for these divs would be display: block, which would cause the text to
be drawn on the next line, anyway.

The actual mistake I made was that I set the text to width: 200px,
which makes the page look correct, until it is resized, at which point
everything stays 200px. I should have set it to max-width of 200px,
which would allow the boxes to grow and shrink as needed. Again, the
solution applied this to the plant class (.info in the solution file),
but I think it works effectively the same to apply it directly to the
text class.