Update styles and html to complete Exercise 05
I missed a couple of things here, though I think my solution looked ok.
I fiddled with margins, padding, and gaps on individual elements when I
could have just added padding and gaps to the entire modal class. I
also continue to make things more complex than they need to be. Instead
of just setting the modal and header to flex (after moving the close
button into that div), and leaving the rest in block display, I decided
to tweak just about everything. The containing div for everything but
the icon is needed, but I didn't need to set the body text as flex, nor
did I need to create a new class to wrap the continue and cancel
buttons. I need to be better about letting the code do what it does and
not re-inventing things.