Edit css to complete Exercise 06
This was a fun one. I initially just started editing, but quickly took
a step back to grab pencil and paper and sketch out what the entire
layout would look like, and give me an idea of what boxes needed to be
where. I think this made it much faster for me to get everything at
least where it needed to be, then it was just a matter of getting it to
look "right".

As in Exercise 03, the only real difference between my code and the
solution is that I changed the left-links and right-links classes to be
flex boxes, rather than changing the ul. I still prefer my method, even
though this appears to be the entire webpage, and those are the only
unordered lists. That might change in the future, though.

I did try to not change the .content to a flex-box, instead using
justify-content on the body to space everything evenly and center the
main content. However, that causes some weirdness when the page shrinks
horizontally, so I had to set it to flex column.

Also, I see now that setting justify-content on the body is unneeded,
as we can set the content to flex: 1, which tells the browser that the
content should fill all the space. I think this again effectively works
the same way either way, but I think I like the solution method more.