Modify html and css to complete Exercise 07
This was another one where I drew out the page before trying to edit
anything. I quickly noticed that the cards would need to be wrapped in
a div, and then that new div and the sidebar would need to be wrapped
in another div. Setting the entire page to flex columns gave the header
and footer layout, but I'd need an inner main body div that could be
set to row, and then another row flex box for the cards.

Having the plan ahead of time really made this go much faster. I got
the main layout down, and then it was just a matter of adjusting
spacing.