Change display to grid and adjust layout
Grid is so much easier than flexbox! At least, for the main layout. I
think the sizes might still require some careful choices, but overall,
it's really just so much simpler to lay things out and adjust them. I'm
also a little impressed with myself that I nailed the measurements for
this task, according to the self-check.