Finalize the grid settings
Not my best work, as I definitely overthought the row settings, which,
apparently aren't even needed? It makes sense, as the browser will just
add them as needed? It's just a bit weird to have all of the
instruction to this point be all about setting both rows and columns
only to have the solution to this have no rows settings. Lesson
learned, though.